rtsp: Don't post AUTimeout check during pause when EOS is nearing
Fix crash issues while repeated pause/resume. The root cause is:
While doing repeated pause-resume operation in RTSP streaming,
some times, the accessunit timeout check is being posted when
the streaming paused, and even after receving the EOS.
Due to this check, after the EOS, the number of packets received
will be zero and the session gets Teardown while the
streaming is paused. Upon resume, it leads to check failure and crash.

Cancel the timeout check if pause is not issued to MyHandler when
nearing EOS.

Change-Id: I897d8e18b0dbc29fc1099e7b3d1b5ccc16426a4d
CRs-Fixed: 636312, 676054